refactor: increase the capacity of the code
need to review to check it right or not
